Extract _LazyModule to dedicated module
