+++ trl/__init__.py
@@ -17,7 +17,7 @@ from importlib.metadata import PackageNotFoundError, version
 from typing import TYPE_CHECKING
 
 from . import _compat
-from .import_utils import _LazyModule
+from ._lazy_module import _LazyModule
 
 
 try:
