+++ trl/rewards/__init__.py
@@ -15,7 +15,7 @@
 import sys
 from typing import TYPE_CHECKING
 
-from ..import_utils import _LazyModule
+from .._lazy_module import _LazyModule
 
 
 _import_structure = {
